backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
boto3==1.28.85